my_contact_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose
python-dotenv
fastapi-limiter
fastapi-cors
cloudinary
fastapi-mail
